Change: Remove function extension template type & move dependency management
