+++ jsonpath4k/src/commonMain/kotlin/at/asitplus/jsonpath/JsonPath.kt
@@ -1,14 +1,26 @@
 package at.asitplus.jsonpath
 
 import at.asitplus.jsonpath.core.JsonPathCompiler
+import at.asitplus.jsonpath.core.JsonPathFilterExpressionType
 import at.asitplus.jsonpath.core.JsonPathFunctionExtension
 import at.asitplus.jsonpath.core.NodeList
+import at.asitplus.jsonpath.core.functionExtensions.countFunctionExtension
+import at.asitplus.jsonpath.core.functionExtensions.lengthFunctionExtension
+import at.asitplus.jsonpath.core.functionExtensions.matchFunctionExtension
+import at.asitplus.jsonpath.core.functionExtensions.searchFunctionExtension
+import at.asitplus.jsonpath.core.functionExtensions.valueFunctionExtension
+import at.asitplus.jsonpath.implementation.AntlrJsonPathCompiler
+import at.asitplus.jsonpath.implementation.AntlrJsonPathCompilerErrorListener
+import io.github.aakira.napier.Napier
 import kotlinx.serialization.json.JsonElement
+import org.antlr.v4.kotlinruntime.BaseErrorListener
+import org.antlr.v4.kotlinruntime.RecognitionException
+import org.antlr.v4.kotlinruntime.Recognizer
 
 class JsonPath(
     jsonPathExpression: String,
-    compiler: JsonPathCompiler = JsonPathDependencyManager.compiler,
-    functionExtensionRetriever: (String) -> JsonPathFunctionExtension<*>? = JsonPathDependencyManager.functionExtensionRepository::getExtension
+    compiler: JsonPathCompiler = defaultCompiler,
+    functionExtensionRetriever: (String) -> JsonPathFunctionExtension? = defaultFunctionExtensionRepository::getExtension
 ) {
     private val query = compiler.compile(
         jsonPath = jsonPathExpression,
@@ -17,5 +29,85 @@ class JsonPath(
 
     fun query(jsonElement: JsonElement): NodeList {
         return query.invoke(jsonElement)
+    }
+
+    companion object {
+        /**
+         * Default json path compiler used when no compiler is explicitly chosen
+         */
+        var defaultCompiler: JsonPathCompiler = AntlrJsonPathCompiler(
+            errorListener = napierAntlrJsonPathCompilerErrorListener,
+        )
+
+        /**
+         * Function extension repository that may be extended with custom functions by the user of this library.
+         */
+        var defaultFunctionExtensionRepository: JsonPathFunctionExtensionRepository = JsonPathFunctionExtensionMapRepository(
+            listOf(
+                lengthFunctionExtension,
+                countFunctionExtension,
+                matchFunctionExtension,
+                searchFunctionExtension,
+                valueFunctionExtension,
+            ).toMap().toMutableMap()
+        )
+    }
+}
+
+
+private val napierAntlrJsonPathCompilerErrorListener by lazy {
+    object : AntlrJsonPathCompilerErrorListener, BaseErrorListener() {
+        override fun unknownFunctionExtension(functionExtensionName: String) {
+            Napier.e {
+                "Unknown JSONPath function extension: \"$functionExtensionName\""
+            }
+        }
+
+        override fun invalidFunctionExtensionForTestExpression(functionExtensionName: String) {
+            Napier.e {
+                "Invalid JSONPath function extension return type for test expression: \"$functionExtensionName\""
+            }
+        }
+
+        override fun invalidFunctionExtensionForComparable(functionExtensionName: String) {
+            Napier.e {
+                "Invalid JSONPath function extension return type for comparable expression: \"$functionExtensionName\""
+            }
+        }
+
+        override fun invalidArglistForFunctionExtension(
+            functionExtensionName: String,
+            functionExtensionImplementation: JsonPathFunctionExtension,
+            coercedArgumentTypes: List<Pair<JsonPathFilterExpressionType?, String>>
+        ) {
+            Napier.e {
+                "Invalid arguments for function extension \"$functionExtensionName\": Expected: <${
+                    functionExtensionImplementation.argumentTypes.joinToString(
+                        ", "
+                    )
+                }>, but received <${
+                    coercedArgumentTypes.map { it.first }.joinToString(", ")
+                }>: <${coercedArgumentTypes.map { it.second }.joinToString(", ")}>"
+            }
+        }
+
+        override fun invalidTestExpression(testContextString: String) {
+            Napier.e {
+                "Invalid test expression: $testContextString"
+            }
+        }
+
+        override fun syntaxError(
+            recognizer: Recognizer<*, *>,
+            offendingSymbol: Any?,
+            line: Int,
+            charPositionInLine: Int,
+            msg: String,
+            e: RecognitionException?
+        ) {
+            Napier.e {
+                "Syntax error $line:$charPositionInLine $msg"
+            }
+        }
     }
 }
