+++ jsonpath4k/src/commonMain/kotlin/at/asitplus/jsonpath/JsonPathFunctionExtensionMapRepository.kt
@@ -3,11 +3,11 @@ package at.asitplus.jsonpath
 import at.asitplus.jsonpath.core.JsonPathFunctionExtension
 
 internal class JsonPathFunctionExtensionMapRepository(
-    private val extensions: MutableMap<String, JsonPathFunctionExtension<*>> = mutableMapOf()
+    private val extensions: MutableMap<String, JsonPathFunctionExtension> = mutableMapOf()
 ) : JsonPathFunctionExtensionRepository {
     override fun addExtension(
         name: String,
-        extension: () -> JsonPathFunctionExtension<*>
+        extension: () -> JsonPathFunctionExtension
     ) {
         extensions[name]?.let {
             throw FunctionExtensionCollisionException(
@@ -17,6 +17,6 @@ internal class JsonPathFunctionExtensionMapRepository(
         extensions[name] = extension()
     }
 
-    override fun getExtension(name: String): JsonPathFunctionExtension<*>? = extensions[name]
-    override fun export(): Map<String, JsonPathFunctionExtension<*>> = extensions.toMap()
+    override fun getExtension(name: String): JsonPathFunctionExtension? = extensions[name]
+    override fun export(): Map<String, JsonPathFunctionExtension> = extensions.toMap()
 }
