+++ jsonpath4k/src/commonMain/kotlin/at/asitplus/jsonpath/JsonPathFunctionExtensionRepository.kt
@@ -10,9 +10,9 @@ interface JsonPathFunctionExtensionRepository {
     /**
      * Implementations should throw FunctionExtensionCollisionException if an extension with that name already exists.
      */
-    fun addExtension(name: String, extension: () -> JsonPathFunctionExtension<*>)
-    fun getExtension(name: String): JsonPathFunctionExtension<*>?
-    fun export(): Map<String, JsonPathFunctionExtension<*>>
+    fun addExtension(name: String, extension: () -> JsonPathFunctionExtension)
+    fun getExtension(name: String): JsonPathFunctionExtension?
+    fun export(): Map<String, JsonPathFunctionExtension>
 }
 
 open class FunctionExtensionCollisionException(message: String) : Exception(message)
