+++ jsonpath4k/src/commonMain/kotlin/at/asitplus/jsonpath/core/JsonPathCompiler.kt
@@ -3,6 +3,6 @@ package at.asitplus.jsonpath.core
 interface JsonPathCompiler {
     fun compile(
         jsonPath: String,
-        functionExtensionRetriever: (String) -> JsonPathFunctionExtension<*>?
+        functionExtensionRetriever: (String) -> JsonPathFunctionExtension?
     ): JsonPathQuery
 }
