+++ jsonpath4k/src/commonTest/kotlin/at/asitplus/jsonpath/DependencyManagementTest.kt
@@ -3,13 +3,12 @@ package at.asitplus.jsonpath
 import at.asitplus.jsonpath.core.JsonPathCompiler
 import at.asitplus.jsonpath.core.JsonPathFunctionExtension
 import at.asitplus.jsonpath.core.JsonPathQuery
-import at.asitplus.jsonpath.core.NodeList
+import at.asitplus.jsonpath.core.JsonPathSelector
 import io.kotest.assertions.throwables.shouldNotThrowAny
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.collections.shouldBeIn
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.nulls.shouldNotBeNull
-import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonNull
 import kotlinx.serialization.json.buildJsonObject
 
@@ -64,17 +63,10 @@ class DependencyManagementTest : FreeSpec({
         val incorrectEmptyQueryCompiler = object : JsonPathCompiler {
             override fun compile(
                 jsonPath: String,
-                functionExtensionRetriever: (String) -> JsonPathFunctionExtension<*>?,
-            ): JsonPathQuery {
-                return object : JsonPathQuery {
-                    override val isSingularQuery: Boolean
-                        get() = true
-
-                    override fun invoke(currentNode: JsonElement, rootNode: JsonElement): NodeList {
-                        return listOf()
-                    }
-                }
-            }
+                functionExtensionRetriever: (String) -> JsonPathFunctionExtension?,
+            ) = JsonPathQuery(
+                selectors = listOf()
+            )
         }
         JsonPathDependencyManager.compiler = incorrectEmptyQueryCompiler
         val emptyQueryResult = JsonPath("$").query(buildJsonObject {})
